+++ ktrack_api/mongo_impl/ktrack_mongo_impl.py
@@ -3,7 +3,7 @@ from mongoengine import connect, DictField
 
 from ktrack_api.ktrack import KtrackIdType
 from ktrack_api.mongo_impl import entities
-from ktrack_api.Exceptions import EntityMissing, EntityNotFoundException
+from ktrack_api.exceptions import EntityMissing, EntityNotFoundException
 from ktrack_api.mongo_impl.entities import NonProjectEntity
 
 
@@ -110,7 +110,7 @@ class KtrackMongoImpl(object):
         entity_candidates = entity_cls.objects(id=entity_id).all()
 
         if len(entity_candidates) == 0:
-            raise EntityNotFoundException(str(entity_id))  # todo would be better to return None instead of exception
+            return None
 
         return _convert_to_dict(entity_candidates[0])