+++ CHANGELOG.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+## 0.5.0 - 2018-08-15
 ### Added
 - Config Manager for unified way to load and validate config files
 - Tests can also run in Nuke and Houdini
+- Added unit and integration tests for ktrack command
 ### Changed
 - Rework Exceptions: reduce throwing and custom exceptions
 - Config Manager is used to load config data